services/user_service/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
loguru>=0.7.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
redis[asyncio]>=5.0.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
